+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All important changes to this project will be documented in this file.
 * Fixed `\n` processing in command `auth ban` for Minecraft servers below version `1.7.6`
 * Added suppression if error `AccessDenied` is raised
 * Made parameter `count` in `clear` command not required when only roles are specified
+* Changed to edit poll message instead of sending a new one
 
 ## [1.3.0](https://github.com/Druzai/Bot_Mc_discord/compare/1.2.3...1.3.0) (2022-05-01)